Refactor: move all files to folder
